pyproject.toml
[tool.poetry]
name = "drf-simple-wallet"
version = "0.1.0"
description = "B2Broker test task"
authors = ["Yast <yastcher@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
djangorestframework-jsonapi = "^7.0.2"
mysqlclient = "^2.2.4"
django-filter = "^24.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.10"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
